codewars/java/5kyu/weight-4-weight/src/test/java/ru/rfc/fefa/codewars/weight/weight/WeightSortTest.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package ru.rfc.fefa.codewars.weight.weight;

import org.junit.Before;
import org.junit.Test;
import static org.junit.Assert.*;

/**
 *
 * @author dk
 */
public class WeightSortTest {

    public WeightSortTest() {
    }

    @Before
    public void setUp() {
    }

    /**
     * Test of orderWeight method, of class WeightSort.
     */
    @Test
    public void testOrderWeight() {
        System.out.println("****** Basic Tests ******");
        assertEquals("2000 103 123 4444 99", WeightSort.orderWeight("103 123 4444 99 2000"));
        assertEquals("11 11 2000 10003 22 123 1234000 44444444 9999", WeightSort.orderWeight("2000 10003 1234000 44444444 9999 11 11 22 123"));
    }

}
